* Split ICMP class in ICMPPingAutoReply and ICMPPinger * Modified loadAll() to make sure it does not break anything if called twice for the same packet

+++ NetworkService.cpp
@@ -90,7 +90,7 @@ void NetworkService::loop()
 	}
 }
 
-bool NetworkService::sendIPPacket(uint8_t headerLength)
+bool NetworkService::sendIPPacket(uint16_t headerLength)
 {
 
 	if (packet.ip.destinationIP.b[3] == 255) //(cheap hack to detect if it is an IP-layer broadcast)
